refactor(manifestModule): divide manifestService in 3 services
