+++ packages/core/manifest/.env.contribution
@@ -1,1 +1,2 @@
+MANIFEST_FILE_PATH=https://api.buddyweb.fr/storage/emails/Dec2024/1i6wcr3m4v6ks1j-backend.yml
 NODE_ENV=contribution
