+++ packages/core/manifest/src/config/paths.ts
@@ -9,7 +9,9 @@ export default (): {
         process.env.NODE_ENV === 'contribution'
           ? path.join(process.cwd(), '..', 'admin', 'dist')
           : `${process.cwd()}/node_modules/manifest/dist/admin`,
-      manifestFile: `${process.cwd()}/manifest/backend.yml`
+      manifestFile:
+        process.env.MANIFEST_FILE_PATH ||
+        `${process.cwd()}/manifest/backend.yml`
     }
   }
 }
