+++ packages/core/manifest/src/crud/guards/is-single.guard.ts
@@ -1,14 +1,14 @@
 import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common'
 import { EntityManifest } from '@repo/types'
-import { ManifestService } from '../../manifest/services/manifest.service'
+import { EntityManifestService } from '../../manifest/services/entity-manifest.service'
 
 @Injectable()
 export class IsSingleGuard implements CanActivate {
-  constructor(private readonly manifestService: ManifestService) {}
+  constructor(private readonly entityManifestService: EntityManifestService) {}
 
   canActivate(context: ExecutionContext): boolean {
     const entityManifest: EntityManifest =
-      this.manifestService.getEntityManifest({
+      this.entityManifestService.getEntityManifest({
         slug: context.getArgs()[0].params.entity
       })
