+++ packages/core/manifest/src/manifest/tests/manifest.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing'
 import { ManifestController } from '../controllers/manifest.controller'
-import { ManifestService } from '../services/manifest.service'
+import { AppManifestService } from '../services/manifest.service'
 import { AuthService } from '../../auth/auth.service'
 
 describe('ManifestController', () => {
@@ -11,7 +11,7 @@ describe('ManifestController', () => {
       controllers: [ManifestController],
       providers: [
         {
-          provide: ManifestService,
+          provide: AppManifestService,
           useValue: {
             getAppManifest: jest.fn(),
             getEntityManifest: jest.fn()
@@ -32,4 +32,15 @@ describe('ManifestController', () => {
   it('should be defined', () => {
     expect(controller).toBeDefined()
   })
+
+  it('should get the app manifest', () => {
+    controller.getAppManifest()
+      
+      expect(controller.manifestService.getAppManifest).toHaveBeenCalled()
+    })
+  })
+
+  it('should get the entity manifest', () => {
+  })
+
 })
