+++ packages/core/manifest/src/manifest/tests/schema.service.spec.ts
@@ -15,4 +15,14 @@ describe('SchemaService', () => {
   it('should be defined', () => {
     expect(service).toBeDefined()
   })
+
+  it('should validate the schema', () => {})
+
+  it('should throw an error if the schema is invalid', () => {})
+
+  describe('custom logic validation', () => {
+    it('should check that all entities in relationships are valid', () => {})
+
+    it('should check that all entities in policies are valid', () => {})
+  })
 })
